refactor: extract webdriver utilities
